requirements-deploy.txt
fastapi
uvicorn
streamlit
tensorflow
pillow
numpy
